Split apart the test "event" handling from the console parsing.
The idea is that the latter can ultimately be used to also handle server
tests.
